user inner modal component
