+++ frontend/src/routes/(root)/(logged)/+layout.svelte
@@ -129,7 +129,7 @@
 		syncTutorialsTodos()
 		loadHubBaseUrl()
 		loadUsedTriggerKinds()
-		loadCriticalAlertsMuted()
+		manageSubscriptions()
 	}
 
 	async function loadUsage() {
@@ -279,6 +279,31 @@
 	}
 
 	setContext('openSearchWithPrefilledText', openSearchModal)
+
+    let enterpriseLoaded = false;
+    let superadminLoaded = false;
+
+    $: {
+        if (enterpriseLoaded && superadminLoaded) {
+            loadCriticalAlertsMuted();
+        }
+    }
+
+    function manageSubscriptions() {
+        const unsubscribeEnterprise = enterpriseLicense.subscribe((value) => {
+            enterpriseLoaded = !!value;
+        });
+
+        const unsubscribeSuperadmin = superadmin.subscribe((value) => {
+            superadminLoaded = !!value;
+        });
+
+        // Clean up subscriptions on component destroy
+        return () => {
+            unsubscribeEnterprise();
+            unsubscribeSuperadmin();
+        };
+    }
 
 	let numUnaknowledgedCriticalAlerts = 0
 	let isCriticalAlertsModalOpen = false
